+++ demo/app/views.py
@@ -14,3 +14,7 @@ def changelist_view(request):
 
 def add_view(request):
     return HttpResponse("This is a dummy add page.")
+
+
+def edit_view(request, object_id):
+    return HttpResponse(f"This is a dummy edit page for object with id '{object_id}'.")